+++ assets/Resources/main/scenes/CreditsScene.cpp
@@ -62,7 +62,7 @@ void CreditsScene::init() {
 void CreditsScene::update(int deltaTime) {
 	currentTime += deltaTime;
 
-	for (int i = 0; i < buttons.size(); i++) buttons[i].update(deltaTime);
+	for (int i = 0; i < (int)buttons.size(); i++) buttons[i].update(deltaTime);
 
 	if (enableControls) {
 		if (Game::instance().getSpecialKey(GLUT_KEY_LEFT) && !latchKeys[GLUT_KEY_LEFT]) {
@@ -84,7 +84,7 @@ void CreditsScene::render() {
 	if (content.size() != 0) {
 		bool first = true;
 		float aux = posText.y;
-		for (int i = 0; i < content.size(); ++i) {
+		for (int i = 0; i < (int)content.size(); ++i) {
 				if (first) {
 					first = false;
 					text.render(content[i], glm::vec2(posText.x, posText.y + 26 * i), 50, color, textAlignment);
